+++ packages/storage/src/namespace_helpers.rs
@@ -1,10 +1,10 @@
-use cosmwasm_std::{ReadonlyStorage, Storage};
+use cosmwasm_std::{ReadonlyStorage, Result, Storage};
 
 pub(crate) fn get_with_prefix<S: ReadonlyStorage>(
     storage: &S,
     namespace: &[u8],
     key: &[u8],
-) -> Option<Vec<u8>> {
+) -> Result<Option<Vec<u8>>> {
     let mut k = namespace.to_vec();
     k.extend_from_slice(key);
     storage.get(&k)
@@ -15,13 +15,17 @@ pub(crate) fn set_with_prefix<S: Storage>(
     namespace: &[u8],
     key: &[u8],
     value: &[u8],
-) {
+) -> Result<()> {
     let mut k = namespace.to_vec();
     k.extend_from_slice(key);
     storage.set(&k, value)
 }
 
-pub(crate) fn remove_with_prefix<S: Storage>(storage: &mut S, namespace: &[u8], key: &[u8]) {
+pub(crate) fn remove_with_prefix<S: Storage>(
+    storage: &mut S,
+    namespace: &[u8],
+    key: &[u8],
+) -> Result<()> {
     let mut k = namespace.to_vec();
     k.extend_from_slice(key);
     storage.remove(&k)
@@ -125,13 +129,13 @@ mod test {
         let prefix = key_prefix(b"foo");
 
         // we use a block scope here to release the &mut before we use it in the next storage
-        set_with_prefix(&mut storage, &prefix, b"bar", b"gotcha");
-        let rfoo = get_with_prefix(&storage, &prefix, b"bar");
+        set_with_prefix(&mut storage, &prefix, b"bar", b"gotcha").unwrap();
+        let rfoo = get_with_prefix(&storage, &prefix, b"bar").unwrap();
         assert_eq!(Some(b"gotcha".to_vec()), rfoo);
 
         // no collisions with other prefixes
         let other_prefix = key_prefix(b"fo");
-        let collision = get_with_prefix(&storage, &other_prefix, b"obar");
+        let collision = get_with_prefix(&storage, &other_prefix, b"obar").unwrap();
         assert_eq!(None, collision);
     }
 }