+++ packages/storage/src/prefix.rs
@@ -1,4 +1,4 @@
-use cosmwasm_std::{ReadonlyStorage, Storage};
+use cosmwasm_std::{ReadonlyStorage, Result, Storage};
 
 use crate::namespace_helpers::{
     get_with_prefix, key_prefix, key_prefix_nested, remove_with_prefix, set_with_prefix,
@@ -41,7 +41,7 @@ impl<'a, T: ReadonlyStorage> ReadonlyPrefixedStorage<'a, T> {
 }
 
 impl<'a, T: ReadonlyStorage> ReadonlyStorage for ReadonlyPrefixedStorage<'a, T> {
-    fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
+    fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
         get_with_prefix(self.storage, &self.prefix, key)
     }
 }
@@ -70,17 +70,17 @@ impl<'a, T: Storage> PrefixedStorage<'a, T> {
 }
 
 impl<'a, T: Storage> ReadonlyStorage for PrefixedStorage<'a, T> {
-    fn get(&self, key: &[u8]) -> Option<Vec<u8>> {
+    fn get(&self, key: &[u8]) -> Result<Option<Vec<u8>>> {
         get_with_prefix(self.storage, &self.prefix, key)
     }
 }
 
 impl<'a, T: Storage> Storage for PrefixedStorage<'a, T> {
-    fn set(&mut self, key: &[u8], value: &[u8]) {
+    fn set(&mut self, key: &[u8], value: &[u8]) -> Result<()> {
         set_with_prefix(self.storage, &self.prefix, key, value)
     }
 
-    fn remove(&mut self, key: &[u8]) {
+    fn remove(&mut self, key: &[u8]) -> Result<()> {
         remove_with_prefix(self.storage, &self.prefix, key)
     }
 }
@@ -96,16 +96,16 @@ mod test {
 
         // we use a block scope here to release the &mut before we use it in the next storage
         let mut foo = PrefixedStorage::new(b"foo", &mut storage);
-        foo.set(b"bar", b"gotcha");
-        assert_eq!(Some(b"gotcha".to_vec()), foo.get(b"bar"));
+        foo.set(b"bar", b"gotcha").unwrap();
+        assert_eq!(Some(b"gotcha".to_vec()), foo.get(b"bar").unwrap());
 
         // try readonly correctly
         let rfoo = ReadonlyPrefixedStorage::new(b"foo", &storage);
-        assert_eq!(Some(b"gotcha".to_vec()), rfoo.get(b"bar"));
+        assert_eq!(Some(b"gotcha".to_vec()), rfoo.get(b"bar").unwrap());
 
         // no collisions with other prefixes
         let fo = ReadonlyPrefixedStorage::new(b"fo", &storage);
-        assert_eq!(None, fo.get(b"obar"));
+        assert_eq!(None, fo.get(b"obar").unwrap());
 
         // Note: explicit scoping is not required, but you must not refer to `foo` anytime after you
         // initialize a different PrefixedStorage. Uncomment this to see errors:
@@ -119,18 +119,18 @@ mod test {
         // set with nested
         let mut foo = PrefixedStorage::new(b"foo", &mut storage);
         let mut bar = PrefixedStorage::new(b"bar", &mut foo);
-        bar.set(b"baz", b"winner");
+        bar.set(b"baz", b"winner").unwrap();
 
         // we can nest them the same encoding with one operation
         let loader = ReadonlyPrefixedStorage::multilevel(&[b"foo", b"bar"], &storage);
-        assert_eq!(Some(b"winner".to_vec()), loader.get(b"baz"));
+        assert_eq!(Some(b"winner".to_vec()), loader.get(b"baz").unwrap());
 
         // set with multilevel
         let mut foobar = PrefixedStorage::multilevel(&[b"foo", b"bar"], &mut storage);
-        foobar.set(b"second", b"time");
+        foobar.set(b"second", b"time").unwrap();
 
         let a = ReadonlyPrefixedStorage::new(b"foo", &storage);
         let b = ReadonlyPrefixedStorage::new(b"bar", &a);
-        assert_eq!(Some(b"time".to_vec()), b.get(b"second"));
+        assert_eq!(Some(b"time".to_vec()), b.get(b"second").unwrap());
     }
 }