+++ packages/vm/src/errors.rs
@@ -89,6 +89,10 @@ pub enum Error {
     },
 }
 
+pub fn make_runtime_err<T>(msg: &'static str) -> Result<T> {
+    RuntimeErr { msg }.fail()
+}
+
 pub type Result<T, E = Error> = core::result::Result<T, E>;
 
 pub trait CacheExt<T: Debug> {